+++ packages/ai-jsx/src/core/completion.tsx
@@ -3,7 +3,6 @@
  * @packageDocumentation
  */
 
-import { ChatCompletionResponseMessage } from 'openai';
 import * as AI from '../index.js';
 import { Node, Component, RenderContext } from '../index.js';
 import { AIJSXError, ErrorCode } from '../core/errors.js';
@@ -12,6 +11,14 @@ import { getEnvVar } from '../lib/util.js';
 import { AnthropicChatModel } from '../lib/anthropic.js';
 import z from 'zod';
 import { zodToJsonSchema } from 'zod-to-json-schema';
+export {
+  UserMessage,
+  SystemMessage,
+  AssistantMessage,
+  FunctionCall,
+  FunctionResponse,
+  ConversationHistory,
+} from './conversation.js';
 
 /**
  * Represents properties passed to a given Large Language Model.
@@ -209,122 +216,6 @@ export function ChatProvider<T extends ModelPropsWithChildren>(
       {children}
     </chatContext.Provider>
   );
-}
-
-/**
- * Provide a System Message to the LLM, for use within a {@link ChatCompletion}.
- *
- * The system message can be used to put the model in character. See https://platform.openai.com/docs/guides/gpt/chat-completions-api for more detail.
- *
- * @example
- * ```tsx
- *    <ChatCompletion>
- *      <SystemMessage>You are a helpful customer service agent.</SystemMessage>
- *    </ChatCompletion>
- * ```
- */
-export function SystemMessage({ children }: { children: Node }) {
-  return children;
-}
-
-/**
- * Provide a User Message to the LLM, for use within a {@link ChatCompletion}.
- *
- * The user message tells the model what the user has said. See https://platform.openai.com/docs/guides/gpt/chat-completions-api for more detail.
- *
- * @example
- * ```tsx
- *    <ChatCompletion>
- *      <UserMessage>I'd like to cancel my account.</UserMessage>
- *    </ChatCompletion>
- *
- *    ==> 'Sorry to hear that. Can you tell me why?
- * ```
- */
-export function UserMessage({ children }: { name?: string; children: Node }) {
-  return children;
-}
-
-/**
- * Provide an Assistant Message to the LLM, for use within a {@link ChatCompletion}.
- *
- * The assistant message tells the model what it has previously said. See https://platform.openai.com/docs/guides/gpt/chat-completions-api for more detail.
- *
- * @example
- * ```tsx
- *    <ChatCompletion>
- *      <UserMessage>I'd like to cancel my account.</UserMessage>
- *      <AssistantMessage>Sorry to hear that. Can you tell me why?</AssistantMessage>
- *      <UserMessage>It's too expensive.</UserMessage>
- *    </ChatCompletion>
- * ```
- *
- *    ==> "Ok, thanks for that feedback. I'll cancel your account."
- */
-export function AssistantMessage({ children }: { children: Node }) {
-  return children;
-}
-
-export function ConversationHistory({ messages }: { messages: ChatCompletionResponseMessage[] }) {
-  return messages.map((message) => {
-    switch (message.role) {
-      case 'system':
-        return <SystemMessage>{message.content}</SystemMessage>;
-      case 'user':
-        return <UserMessage>{message.content}</UserMessage>;
-      case 'assistant':
-        return <AssistantMessage>{message.content}</AssistantMessage>;
-      case 'function':
-        return (
-          <FunctionCall name={message.function_call!.name!} args={JSON.parse(message.function_call!.arguments!)} />
-        );
-    }
-  });
-}
-
-/**
- * Provide a function call to the LLM, for use within a {@link ChatCompletion}.
- *
- * The function call tells the model that a function was previously invoked by the model. See https://platform.openai.com/docs/guides/gpt/chat-completions-api for more detail.
- * When the model returns a function call, @{link ChatCompletion} returns a @{link FunctionCall} component.
- *
- * @example
- * ```tsx
- *    <ChatCompletion>
- *      <UserMessage>What is 258 * 322?</UserMessage>
- *      <FunctionCall name="evaluate_math" args={expression: "258 * 322"} />
- *      <FunctionResponse name="evaluate_math">83076</FunctionResponse>
- *    </ChatCompletion>
- *
- *    ==> "That would be 83,076."
- * ```
- */
-export function FunctionCall({ name, args }: { name: string; args: Record<string, string | number | boolean | null> }) {
-  return `Call function ${name} with ${JSON.stringify(args)}`;
-}
-
-/**
- * Renders to the output of a previous {@link FunctionCall} component, for use within a {@link ChatCompletion}.
- *
- * See https://platform.openai.com/docs/guides/gpt/chat-completions-api for more detail.
- *
- * @example
- * ```tsx
- *    <ChatCompletion>
- *      <UserMessage>What is 258 * 322?</UserMessage>
- *      <FunctionCall name="evaluate_math" args={expression: "258 * 322"} />
- *      <FunctionResponse name="evaluate_math">83076</FunctionResponse>
- *    </ChatCompletion>
- *
- *    ==> "That would be 83,076."
- * ```
- */
-export async function FunctionResponse(
-  { name, children }: { name: string; children: Node },
-  { render }: AI.ComponentContext
-) {
-  const output = await render(children);
-  return `function ${name} returns ${output}`;
 }
 
 /**