+++ packages/ai-jsx/src/lib/openai.tsx
@@ -19,26 +19,23 @@ import {
 } from 'openai-edge';
 import { Merge, MergeExclusive } from 'type-fest';
 import {
-  AssistantMessage,
   ChatProvider,
   CompletionProvider,
-  FunctionCall,
   FunctionDefinition,
-  FunctionResponse,
   ModelProps,
   ModelPropsWithChildren,
-  SystemMessage,
-  UserMessage,
   getParametersSchema,
 } from '../core/completion.js';
+import { AssistantMessage, FunctionCall, renderToConversation } from '../core/conversation.js';
 import { AIJSXError, ErrorCode, HttpError } from '../core/errors.js';
 import { Image, ImageGenPropsWithChildren } from '../core/image-gen.js';
 import { Logger } from '../core/log.js';
 import * as AI from '../index.js';
-import { Node, PropsOfComponent } from '../index.js';
+import { Node } from '../index.js';
 import { ChatOrCompletionModelOrBoth } from './model.js';
 import { getEnvVar, patchedUntruncateJson } from './util.js';
 import { CreateChatCompletionRequest } from 'openai';
+import { debug } from '../core/debug.js';
 
 // https://platform.openai.com/docs/models/model-endpoint-compatibility
 type ValidCompletionModel =
@@ -296,21 +293,13 @@ export async function* OpenAIChatModel(
       {
         functionDefinitions: Record<string, FunctionDefinition>;
         forcedFunction: string;
-
-        /**
-         * Internal only.
-         *
-         * If this flag is passed, this component will stream only the model Function Call. This is useful if the only thing you want is the function call, such as if you're using the function call to force the model to output JSON matching a schema.
-         */
-        experimental_streamFunctionCallOnly: boolean;
       },
       {
         functionDefinitions?: never;
         forcedFunction?: never;
-        experimental_streamFunctionCallOnly?: never;
       }
     >,
-  { render, getContext, logger }: AI.ComponentContext
+  { render, getContext, logger, memo }: AI.ComponentContext
 ): AI.RenderableStream {
   if (props.functionDefinitions) {
     if (!chatModelSupportsFunctions(props.model)) {
@@ -321,14 +310,6 @@ export async function* OpenAIChatModel(
       );
     }
   }
-  // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-  else if (props.experimental_streamFunctionCallOnly) {
-    throw new AIJSXError(
-      'The experimental_streamFunctionCallOnly flag can only be passed when function definitions are also passed.',
-      ErrorCode.ChatCompletionBadInput,
-      'user'
-    );
-  }
 
   if (props.forcedFunction && !Object.keys(props.functionDefinitions).includes(props.forcedFunction)) {
     throw new AIJSXError(
@@ -338,64 +319,51 @@ export async function* OpenAIChatModel(
     );
   }
 
-  const messageElements = await render(props.children, {
-    stop: (e) =>
-      e.tag == SystemMessage ||
-      e.tag == UserMessage ||
-      e.tag == AssistantMessage ||
-      e.tag == FunctionCall ||
-      e.tag == FunctionResponse,
-  });
-  if (!props.experimental_streamFunctionCallOnly) {
-    yield AI.AppendOnlyStream;
-  }
+  yield AI.AppendOnlyStream;
+
+  const conversationMessages = await renderToConversation(props.children, render);
+  logger.debug({ messages: conversationMessages.map((m) => debug(m.element, true)) }, 'Got input conversation');
   const messages: ChatCompletionRequestMessage[] = await Promise.all(
-    messageElements.filter(AI.isElement).map(async (message) => {
-      switch (message.tag) {
-        case SystemMessage:
+    conversationMessages.map(async (message) => {
+      switch (message.type) {
+        case 'system':
           return {
             role: 'system',
-            content: await render(message),
+            content: await render(message.element),
           };
-        case UserMessage:
+        case 'user':
           return {
             role: 'user',
-            content: await render(message),
-            name: (message.props as PropsOfComponent<typeof UserMessage>).name,
+            content: await render(message.element),
+            name: message.element.props.name,
           };
-        case AssistantMessage:
+        case 'assistant':
           return {
             role: 'assistant',
-            content: await render(message),
+            content: await render(message.element),
           };
-        case FunctionCall:
+        case 'functionCall':
           return {
             role: 'assistant',
             content: '',
             function_call: {
-              name: message.props.name,
-              arguments: JSON.stringify(message.props.args),
+              name: message.element.props.name,
+              arguments: JSON.stringify(message.element.props.args),
             },
           };
-        case FunctionResponse:
+        case 'functionResponse':
           return {
             role: 'function',
-            name: message.props.name,
-            content: await render(message.props.children),
+            name: message.element.props.name,
+            content: await render(message.element.props.children),
           };
-        default:
-          throw new AIJSXError(
-            `ChatCompletion's prompts must be SystemMessage, UserMessage, AssistantMessage, FunctionCall, or FunctionResponse but this child was ${message.tag.name}`,
-            ErrorCode.ChatCompletionUnexpectedChild,
-            'internal'
-          );
       }
     })
   );
 
   if (!messages.length) {
     throw new AIJSXError(
-      "ChatCompletion must have at least child that's a SystemMessage, UserMessage, AssistantMessage, FunctionCall, or FunctionResponse, but no such children were found.",
+      "ChatCompletion must have at least one child that's a SystemMessage, UserMessage, AssistantMessage, FunctionCall, or FunctionResponse, but no such children were found.",
       ErrorCode.ChatCompletionMissingChildren,
       'user'
     );
@@ -438,59 +406,104 @@ export async function* OpenAIChatModel(
     }
   >;
 
-  const currentMessage = { content: undefined, function_call: undefined } as Partial<ChatCompletionResponseMessage>;
-  let finishReason: string | undefined = undefined;
-  for await (const deltaMessage of openAiEventsToJson<ChatCompletionDelta>(
-    asyncIteratorOfFetchStream(chatResponse.body!.getReader())
-  )) {
-    logger.trace({ deltaMessage }, 'Got delta message');
-    finishReason = finishReason ?? deltaMessage.choices[0].finish_reason;
-    const delta = deltaMessage.choices[0].delta;
-    if (delta.role) {
-      currentMessage.role = deltaMessage.choices[0].delta.role;
+  const iterator = openAiEventsToJson<ChatCompletionDelta>(asyncIteratorOfFetchStream(chatResponse.body!.getReader()))[
+    Symbol.asyncIterator
+  ]();
+
+  // We have a single response iterator, but we'll wrap tokens _within_ the structure of <AssistantMessage> or <FunctionCall>
+  // components. This:
+  //  - Allows our stream to be append-only and therefore eagerly rendered in append-only contexts.
+  //  - Preserves the output structure to allow callers to extract/separate <AssistantMessage> and <FunctionCall> messages.
+  //  - Allows the intermediate states of the stream to include "partial" <FunctionCall> elements with healed JSON.
+  //
+  // This requires some gymnastics because several components will share a single iterator that can only be consumed once.
+  // That is, the logical loop execution is spread over multiple functions (closures over the shared iterator).
+  async function advance(): Promise<Partial<ChatCompletionResponseMessage> | null> {
+    const next = await iterator.next();
+    if (next.done) {
+      return null;
     }
-    if (delta.content) {
-      currentMessage.content = currentMessage.content ?? '';
-      currentMessage.content += delta.content;
-      if (!props.experimental_streamFunctionCallOnly) {
-        yield delta.content;
-      }
+
+    logger.trace({ deltaMessage: next.value }, 'Got delta message');
+    return next.value.choices[0].delta;
+  }
+
+  let isAssistant = false;
+  let delta = await advance();
+  while (delta !== null) {
+    if (delta.role === 'assistant') {
+      isAssistant = true;
     }
-    if (delta.function_call) {
-      currentMessage.function_call = currentMessage.function_call ?? { name: '', arguments: '' };
-      if (delta.function_call.name) {
-        currentMessage.function_call.name += delta.function_call.name;
-      }
-      if (delta.function_call.arguments) {
-        currentMessage.function_call.arguments += delta.function_call.arguments;
-      }
-      if (props.experimental_streamFunctionCallOnly) {
-        yield JSON.stringify({
-          ...currentMessage.function_call,
-          // We actually want the argument to be '{}' if it's empty, not '""'.
-          // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
-          arguments: JSON.parse(patchedUntruncateJson(currentMessage.function_call.arguments || '{}')),
-        });
-      }
+
+    if (isAssistant && delta.content) {
+      // Memoize the stream to ensure it renders only once.
+      const assistantStream = memo(
+        (async function* (): AI.RenderableStream {
+          yield AI.AppendOnlyStream;
+
+          while (delta !== null) {
+            if (delta.content) {
+              yield delta.content;
+            }
+            if (delta.function_call) {
+              break;
+            }
+            delta = await advance();
+          }
+
+          return AI.AppendOnlyStream;
+        })()
+      );
+      yield <AssistantMessage>{assistantStream}</AssistantMessage>;
+
+      // Ensure the assistantStream is flushed by rendering it.
+      await render(assistantStream);
+    }
+
+    // TS doesn't realize that the assistantStream closure can make `delta` be `null`.
+    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+    if (delta?.function_call) {
+      // Memoize the stream to ensure it renders only once.
+      const functionCallStream = memo(
+        (async function* () {
+          let name = '';
+          let argsJson = '';
+          while (delta != null) {
+            if (!delta.function_call) {
+              break;
+            }
+
+            if (delta.function_call.name) {
+              name += delta.function_call.name;
+            }
+
+            if (delta.function_call.arguments) {
+              argsJson += delta.function_call.arguments;
+            }
+
+            yield <FunctionCall partial name={name} args={JSON.parse(patchedUntruncateJson(argsJson || '{}'))} />;
+
+            delta = await advance();
+          }
+
+          return <FunctionCall name={name} args={JSON.parse(argsJson || '{}')} />;
+        })()
+      );
+      yield functionCallStream;
+
+      // Ensure the functionCallStream is flushed by rendering it.
+      await render(functionCallStream);
+    }
+
+    // TS doesn't realize that the functionCallStream closure can make `delta` be `null`.
+    // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
+    if (delta !== null) {
+      delta = await advance();
     }
   }
 
-  logger.debug({ message: currentMessage }, 'Finished createChatCompletion');
+  logger.debug('Finished createChatCompletion');
 
-  if (props.experimental_streamFunctionCallOnly) {
-    return JSON.stringify({
-      ...currentMessage.function_call,
-      arguments: JSON.parse(currentMessage.function_call?.arguments ?? '{}'),
-    });
-  }
-  if (currentMessage.function_call) {
-    yield (
-      <FunctionCall
-        name={currentMessage.function_call.name ?? ''}
-        args={JSON.parse(currentMessage.function_call.arguments ?? '{}')}
-      />
-    );
-  }
   return AI.AppendOnlyStream;
 }